debug...

+++ src/vehicle_system/autoaim/armor_detector/include/armor_tracker/armor_tracker.hpp
@@ -38,7 +38,7 @@ namespace armor_detector
         Rect rect;
         Point2f apex2d[4];
         RotatedRect rrect;
-        cv::Point2d center2d;
+        cv::Point2f center2d;
         TargetType type;
         bool is_front;
         double rangle;
